Final version for grading

+++ src/main/java/com/pauldok/cruduserbase/repository/UserBaseRepository.java
@@ -3,5 +3,5 @@ package com.pauldok.cruduserbase.repository;
 import com.pauldok.cruduserbase.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserBaseRepository extends JpaRepository<User, Long>{
+public interface UserBaseRepository extends JpaRepository<User, Integer> {
 }
